Revert "refactor(socket): wip, extract data responsibilities to provider"
This reverts commit f1ce0d4642b72fdf5d76636b9329d871157d6960.
