+++ server/src/controllers/eventController.js
@@ -1,7 +1,5 @@
-const data = global.eventsProvider.getData();
-const updateDb = async (newData) => {
-  await global.eventsProvider.updateData(newData);
-};
+// get database
+import { db, data } from '../app.js';
 
 // Create controller for GET request to 'event'
 // Returns ACK message
@@ -19,7 +17,7 @@ export const postEvent = async (req, res) => {
   // TODO: validate data
   try {
     data.event = { ...data.event, ...req.body };
-    await updateDb(data.event);
+    await db.write();
     res.sendStatus(200);
   } catch (error) {
     res.status(400).send(error);
